refactor(credit-check): consolidate creation to customer detail page
Credit-check list at /credit-checks becomes a read-only queue/dashboard.
All credit-check creation (OCR book bank, salary slip, bank statement,
risk score, approve/reject) moves to the customer detail credit tab via
a shared dialog.

- extract creation state + mutations into useCreditCheckCreate hook
- wrap CreditCheckCreateModal in CreditCheckCreateDialog (stateful) that
  takes { open, onClose, preselectedCustomer, onCreated }
- CustomerDetailPage credit tab: replace simple inline upload form with
  a "+ ตรวจเครดิตใหม่" button opening the dialog (full OCR/risk flow)
- CustomerDetailPage: read ?tab=<name> from URL, sync activeTab both ways
- CreditChecksPage strips ~400 LOC of creation state + modal + OCR
  handlers; row click navigates to /customers/:id?tab=credit instead of
  expanding inline
- delete CreditCheckDetail.tsx (expand panel no longer used — full detail
  is on the customer page tab)
- keep on queue page: filters, table, Excel export, override modal,
  inline AI analyze (1-click op)

Net: -496 LOC (+70 / -566).

Type check: passed. Lint: 0 errors.
